Declare the bundled Issue and PR list-detail couplings instead of hand-wiring them
The issue and pull-request screens have always cleared their detail request
when the list selection moved, and that rule lived in two nearly identical
reducer functions. It is now stated once, in each screen's descriptor, as the
same master-detail relationship a user-authored screen would declare: an output
selection port, an input subject port carrying the same versioned type, and a
relationship that follows immediately and clears when the selection empties.

The reducers ask the descriptor rather than naming ports themselves, so the
declaration stays the only statement of which panels are coupled, and the same
propagation engine carries both a shipped screen and a defined one. The actions
and errors screens declare no coupling, because their detail panes do not follow
their list selection, and that absence is now visible in the descriptor rather
than implied by the absence of glue.

The subject a list publishes is its selected issue or pull-request number
rather than a row index, which is what the coupling actually means; at every
call site the list is unchanged across the transition, so the two agree exactly
and the existing issue and pull-request suites are the parity evidence.

Refs #385
